% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{residuals.nuglmm}
\alias{residuals.nuglmm}
\title{Compute residuals for a nuglmm object}
\usage{
\method{residuals}{nuglmm}(object, type = c("response", "pearson", "PIT"),
  ...)
}
\arguments{
\item{object}{a \dQuote{nuglmm} object}

\item{type}{(character) residual type}

\item{\dots}{ignored, for method compatibility}
}
\description{
Compute residuals for a nuglmm object
}
