% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuglmm.R
\name{nuglmm}
\alias{nuglmm}
\title{Generalized LMM for multivariate abundance data}
\usage{
nuglmm(formula, data = NULL, offset = NULL, weights = NULL, ...)
}
\arguments{
\item{formula}{model specification}

\item{data}{dataframe}

\item{\dots}{optional additional arguments. Currently none are used in any
methods.}
}
\value{
a fitted nuglmm model.
}
\description{
Fits nuglmm models to multiavriate abundance data
}
\examples{
data(Tasmania, package = "mvabund")
m1 = nuglmm(abund ~ treatment * block, data=Tasmania, family="poisson")
m2 = nuglmm(abund ~ treatment * block, data=Tasmania, family="nbinom2")
anova(m1, m2)
}
\keyword{models}
